fix: update fsd structure
