--- .gitignore
@@ -1,7 +1,0 @@
-.vscode/
-.idea/
-node_modules/
-dist/
-.env
-.env.mock
-.env.api
