+++ frontend/src/app/config/router.ts
@@ -1,13 +1,13 @@
 import type { RouteObject } from 'react-router';
 import { createBrowserRouter, replace } from 'react-router';
-import { LazyOfferPage } from '@/pages/OfferPage';
+import { LazyOfferPage } from '@/pages/offer';
 import { PATHS } from '@/shared/consts';
-import { BaseLayout } from '../layout/BaseLayout';
+import { Layout } from '../layout/layout';
 
 const routes: RouteObject[] = [
   {
     path: PATHS.base,
-    Component: BaseLayout,
+    Component: Layout,
     children: [
       {
         index: true,
