+++ frontend/src/app/layout/layout.tsx
@@ -1,0 +1,17 @@
+import { Layout as AntLayout, Skeleton } from 'antd';
+import { Suspense } from 'react';
+import { Outlet } from 'react-router';
+import { Header } from './Header';
+
+export const Layout = () => {
+  return (
+    <AntLayout>
+      <Header />
+      <AntLayout.Content>
+        <Suspense fallback={<Skeleton />}>
+          <Outlet />
+        </Suspense>
+      </AntLayout.Content>
+    </AntLayout>
+  );
+};
