+++ frontend/src/features/flight-filters/index.ts
@@ -1,1 +1,3 @@
-export { FlightFilters } from './ui/FlightFilters';
+export { FlightFilters } from './ui/flight-filters';
+export { useFlightFilters } from './model/use-flight-filters';
+export type { FlightFiltersState } from './model/types';
