+++ frontend/src/features/launch-experiment/index.ts
@@ -1,5 +1,5 @@
-export { LaunchExperimentProvider } from './ui/LaunchExperimentProvider';
-export { useLaunchExperiment } from './lib/useLaunchExperiment';
-export { trackExperimentEvent } from './lib/trackExperiment';
-export { Goal } from './consts/experimentGoals';
-export type { GoalName } from './consts/experimentGoals';
+export { LaunchExperimentProvider } from './ui/launch-experiment-provider';
+export { useLaunchExperiment } from './model/use-launch-experiment';
+export { trackExperimentEvent } from './model/track-experiment';
+export { Goal } from './model/experiment-goals';
+export type { GoalName } from './model/experiment-goals';
