+++ frontend/src/features/launch-experiment/model/get-experiment-variant.ts
