+++ frontend/src/features/launch-experiment/model/use-launch-experiment.ts
@@ -1,0 +1,16 @@
+import type { ExperimentVariant } from './get-experiment-variant';
+import { createContext, useContext } from 'react';
+
+type LaunchExperimentContextType = ExperimentVariant | null;
+
+export const LaunchExperimentContext = createContext<LaunchExperimentContextType>(null);
+
+export const useLaunchExperiment = () => {
+  const context = useContext(LaunchExperimentContext);
+
+  if (context === null) {
+    throw new Error('Контекст для проведения эксперимента не установлен');
+  }
+
+  return context;
+};
