--- frontend/src/features/search-form/consts/defaults.ts
@@ -1,8 +1,0 @@
-import type { PassengersState } from '../types/searchForm';
-
-export const DEFAULT_PASSENGERS: PassengersState = {
-  adults: 1,
-  children: 0,
-  toddler: 0,
-  animals: 0,
-};
