+++ frontend/src/features/search-form/index.ts
@@ -1,1 +1,3 @@
-export { SearchForm } from './ui/SearchForm';
+export { SearchForm } from './ui/search-form';
+export { useSearchForm } from './model/use-search-form';
+export type { AirportOption, PassengersState, ServiceClass, TripType } from './model/types';
