+++ frontend/src/features/search-form/model/defaults.ts
@@ -1,0 +1,8 @@
+import type { PassengersState } from './types';
+
+export const DEFAULT_PASSENGERS: PassengersState = {
+  adults: 1,
+  children: 0,
+  toddler: 0,
+  animals: 0,
+};
