+++ frontend/src/features/search-form/model/map-airport-to-option.ts
@@ -1,0 +1,8 @@
+import type { AirportOption } from './types';
+import type { AirportDto } from '@/shared/types';
+
+export const mapAirportToOption = (a: AirportDto): AirportOption => ({
+  value: a.city,
+  label: `${a.city} — ${a.airport}`,
+  airportId: a.id,
+});
