+++ frontend/src/features/search-form/model/types.ts
