--- frontend/src/features/search-form/utils/getPassengerLabel.ts
@@ -1,8 +1,0 @@
-import type { PassengersState } from '../types/searchForm';
-
-export const getPassengerLabel = (p: PassengersState): string => {
-  const total = p.adults + p.children + p.toddler;
-  if (total === 1) return '1 пассажир';
-  if (total >= 2 && total <= 4) return `${total} пассажира`;
-  return `${total} пассажиров`;
-};
