--- frontend/src/features/search-form/utils/mapAirportToOption.ts
@@ -1,8 +1,0 @@
-import type { AirportOption } from '../types/searchForm';
-import type { AirportDto } from '@/shared/types';
-
-export const mapAirportToOption = (a: AirportDto): AirportOption => ({
-  value: a.city,
-  label: `${a.city} — ${a.airport}`,
-  airportId: a.id,
-});
