--- frontend/src/pages/OfferPage/index.ts
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const LazyOfferPage = lazy(() => import('./ui/OfferPage'));
