+++ frontend/src/pages/offer/index.ts
@@ -1,0 +1,3 @@
+import { lazy } from 'react';
+
+export const LazyOfferPage = lazy(() => import('./ui/offer-page'));
