+++ frontend/src/pages/offer/ui/offer-page.module.css
